requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-widget-tweaks==1.4.5
dnspython==1.16.0
eventlet==0.25.1
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.7
monotonic==1.5
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
python-decouple==3.3
python-dotenv==0.11.0
python-http-client==3.2.5
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==6.1.2
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
whitenoise==5.0.1
zipp==3.0.0
